[FIX] functional tests - fix expected values

+++ leaspy/inputs/data/dataset.py
@@ -33,15 +33,13 @@ class Dataset:
         x_len = [len(_.timepoints) for _ in data]
         channels = data.dimension
         values = torch.zeros((batch_size, max(x_len), channels), dtype=torch.float32)
-        # mask = torch.zeros((batch_size, max(x_len), channels), dtype=torch.float32)
         # padding_mask = np.zeros((batch_size, max(x_len), channels))
-
         padding_mask = torch.zeros((batch_size, max(x_len), channels), dtype=torch.float32)
 
         # TODO missing values in mask ?
 
         for i, d in enumerate(x_len):
-            indiv_values = torch.tensor(data[i].observations)
+            indiv_values = torch.tensor(data[i].observations, dtype=torch.float32)
             values[i, 0:d, :] = indiv_values
             padding_mask[i, 0:d, :] = 1
 
@@ -50,7 +48,7 @@ class Dataset:
         mask = padding_mask * mask_missingvalues
 
         # values[np.array(1 - mask_missingvalues, dtype=bool)] = 0 # Set values of missing values to 0.0
-        values[torch.isnan(values)] = 0 # Set values of missing values to 0.0
+        values[torch.isnan(values)] = 0  # Set values of missing values to 0.0
 
         self.n_individuals = batch_size
         self.max_observations = max(x_len)
@@ -63,7 +61,7 @@ class Dataset:
         self.n_observations = int(mask.sum().item())
 
     def _construct_timepoints(self, data):
-        self.timepoints = torch.zeros([self.n_individuals, self.max_observations])
+        self.timepoints = torch.zeros([self.n_individuals, self.max_observations], dtype=torch.float32)
         x_len = [len(_.timepoints) for _ in data]
         for i, d in enumerate(x_len):
             self.timepoints[i, 0:d] = torch.tensor(data[i].timepoints, dtype=torch.float32)
@@ -79,7 +77,7 @@ class Dataset:
         # mask = self.mask[i].clone().cpu().detach().numpy()[:values.shape[0],:]
         mask = self.mask[i].clone().cpu().detach()[:values.shape[0], :]
         # mask[mask==0] = np.nan
-        mask[mask==0] = float('NaN')
+        mask[mask == 0] = float('NaN')
         values_with_na = values * mask
         return values_with_na
